+++ CHANGELOG.md
@@ -50,4 +50,5 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Issue #36: Wrong curve drawing
 - Issue #38: Trains hanging in DoubleSlipSwitch
 - Issue #41: Train stops at central cross block
-- Issue #62: Exiting train disapper from map panel 
+- Issue #62: Exiting train disappears from map panel
+- Issue #61: Wrong signal lock at train crossing section